Remove dependency of material name on texture dir
Previously the hash in the material name depended on the full
path to each texture, which could vary across machines.

By passing the texture dir separately, it should now depend only
on the texture CRCs, so everyone who rips the same game will get
the same material names.

Also fixes a bug when the CRC is 0.